moving CIDict to avoid circular imports
